Rename det1024 sources and tests to make room for the 512 variant
Give the det1024 files the same explicit parameter suffix the forthcoming
det512 files will use:

  - deterministic.c -> deterministic1024.c
  - tests/test_deterministic.c -> tests/test_deterministic1024.c
  - tests/test_deterministic_kat.h -> tests/test_deterministic1024_kat.h

The Makefile targets, the CI workflow step, and the config.h reference are
updated to match, and the renamed test object rule gains its previously
missing dependency on the KAT header. No functional change.
