+++ .github/workflows/test.yml
@@ -28,8 +28,8 @@ jobs:
           go-version-file: go.mod
       - name: Build C tests
         run: make
-      - name: Deterministic KAT tests
-        run: ./tests/test_deterministic
+      - name: Deterministic-1024 KAT tests
+        run: ./tests/test_deterministic1024
       - name: Falcon tests
         run: ./tests/test_falcon
       - name: Speed smoke test
