+++ deterministic1024.c
